Create README.md
I planned on working on this from 8 AM until 2 PM and then early this morning we were told we had a team building activity that no one was notified of prior. This caused me to have to rush through the code and because of the loss of time, I wasn't even able to get the RecyclerViewAdapter class for the animal list made let alone the rest. Hopefully you understand. I plan on revisiting this.